test: move test utilities outside `src` directory
