+++ src/solution.test.ts
@@ -2,7 +2,7 @@ import { readFile } from "node:fs/promises";
 import path from "node:path";
 import { afterAll, expect, test, vi } from "vitest";
 import { testCppSolution } from "./internal/solution.js";
-import { createTempFs, removeAllTempFs } from "./internal/utils/temp-fs.js";
+import { createTempFs, removeAllTempFs } from "../test/temp-fs.js";
 import { type TestResult, testSolutions } from "./solution.js";
 
 vi.mock("./internal/solution.js", () => ({
